Wrap only Timeline in redux
